[Refactor/#81] Refactor: Post 타입을 `@/lib/utils/post`에서 가져오도록 수정
